chore(ava/data): rename dataAnalysis to analysis
