+++ packages/ava/src/data/analysis/index.ts
